time

+++ srcs/ping.c
@@ -93,8 +93,8 @@ static int	send_loop(t_ping *ping, int sock)
   			{ 
 		*/
 				gettimeofday(&aft, NULL);
-  				printf("%zd bytes from %s (%s): icmp_seq=%d ttl=%d time=%.2f ms\n",  
-  					recv_bytes, ping->dest_name, ping->dest_ip, ping->msg_count, PING_TTL, (float)(aft.tv_usec - bef.tv_usec) / 1000); 
+  				printf("%d bytes from %s (%s): icmp_seq=%d ttl=%d time=%.2f ms\n",  
+  					(int)recv_bytes, ping->dest_name, ping->dest_ip, ping->msg_count, PING_TTL, (float)(aft.tv_usec - bef.tv_usec) / 1000); 
   				ping->msg_recv_count++; 
 				ping->total_stime = (ping->total_stime == -1 ? 0 : ping->total_stime + (aft.tv_usec - bef.tv_usec));
   			//}
